% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.code-package.R
\docType{package}
\name{aws.code-package}
\alias{aws.code}
\alias{aws.code-package}
\title{aws.code}
\description{
AWS CodeCommit/CodeDeploy/CodePipeline Client Package
}
\details{
A simple client package for the Amazon Web Services (AWS) CodeCommit/CodeDeploy/CodePipeline REST APIs.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}

