% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{awscodeHTTP}
\alias{awscodeHTTP}
\title{API Requests}
\usage{
awscodeHTTP(action, query = list(), headers = list(), body = list(),
  service = c("CodeCommit", "CodeDeploy", "CodePipeline"),
  verbose = getOption("verbose", FALSE),
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"), key = NULL,
  secret = NULL, session_token = NULL, version = "2015-04-13", ...)
}
\arguments{
\item{action}{A character string specifying an API action.}

\item{query}{A named list of query string parameters.}

\item{headers}{A list of headers to pass to the HTTP request.}

\item{body}{A list of body (JSON) arguments.}

\item{service}{A character string specifying one of \dQuote{CodeCommit}, \dQuote{CodeDeploy}, or \dQuote{CodePipeline}.}

\item{verbose}{A logical indicating whether to be verbose. Default is given by \code{options("verbose")}.}

\item{region}{A character string containing the AWS region. If missing, defaults to \dQuote{us-east-1}.}

\item{key}{A character string containing an AWS Access Key ID. See \code{\link[aws.signature]{locate_credentials}}.}

\item{secret}{A character string containing an AWS Secret Access Key. See \code{\link[aws.signature]{locate_credentials}}.}

\item{session_token}{A character string containing an AWS Session Token. See \code{\link[aws.signature]{locate_credentials}}.}

\item{version}{A character string specifying an API version. Default is \dQuote{2015-04-13}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{POST}}.}
}
\value{
A list
}
\description{
CodeCommit/CodeDeploy/CodePipeline HTTP Requests
}
