% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeCommit_branches.R
\name{get_branch}
\alias{get_branch}
\alias{create_branch}
\alias{get_default_branch}
\alias{set_default_branch}
\title{CodeCommit Branches}
\usage{
get_branch(branch, repo, token, ...)

create_branch(branch, commit, repo, ...)

get_default_branch(repo, branch, ...)

set_default_branch(repo, branch, ...)
}
\arguments{
\item{branch}{A character string specifying a branch name, or an object of class \dQuote{aws_repo_branch}.}

\item{repo}{A character string specifying a repository name, or an object of class \dQuote{aws_repo}.}

\item{token}{A pagination token.}

\item{\dots}{Additional arguments passed to \code{\link{awscodeHTTP}}.}

\item{commit}{A character string specifying a commitId (SHA), or an object of class \dQuote{aws_repo_commit}.}
}
\value{
A list.
}
\description{
Get, create, delete CodeCommit repo branches
}
