% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeCommit_commit.R
\name{get_commit}
\alias{get_commit}
\title{CodeCommit Commits}
\usage{
get_commit(commit, repo, ...)
}
\arguments{
\item{commit}{A character string specifying a commitId (SHA), or an object of class \dQuote{aws_repo_commit}.}

\item{repo}{A character string specifying a repository name, or an object of class \dQuote{aws_repo}.}

\item{\dots}{Additional arguments passed to \code{\link{awscodeHTTP}}.}
}
\value{
A list.
}
\description{
Get CodeCommit repo commit
}

