% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeCommit_repos.R
\name{get_repo}
\alias{create_repo}
\alias{delete_repo}
\alias{get_repo}
\alias{update_repo}
\title{CodeCommit Repositories}
\usage{
get_repo(repo, sort, order, token, ...)

create_repo(name, description, ...)

update_repo(repo, name, description, ...)

delete_repo(repo, ...)
}
\arguments{
\item{repo}{A character string specifying a repository name, or an object of class \dQuote{aws_repo}.}

\item{sort}{One of \dQuote{repositoryName} or \dQuote{lastModifiedDate}.}

\item{order}{One of \dQuote{ascending} or \dQuote{descending}.}

\item{token}{A pagination token.}

\item{name}{A character string specifying a repository name.}

\item{description}{A character string specifying a repository description.}

\item{commit}{A character string specifying a commitId (SHA), or an object of class \dQuote{aws_repo_commit}.}

\item{\dots}{Additional arguments passed to \code{\link{awscodeHTTP}}.}
}
\value{
A list.
}
\description{
Get, create, update, delete CodeCommit repositories
}

