% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeCommit_triggers.R
\name{get_repo_triggers}
\alias{get_repo_triggers}
\alias{make_trigger}
\alias{set_repo_triggers}
\title{CodeCommit Repo Triggers}
\usage{
get_repo_triggers(repo, ...)

set_repo_triggers(repo, triggers = list(), ...)

make_trigger(branch, name, arn, events = "all", custom_data)
}
\arguments{
\item{repo}{A character string specifying a repository name, or an object of class \dQuote{aws_repo}.}

\item{triggers}{A list of objects of class \dQuote{aws_repo_trigger}, as created by \code{make_trigger}.}

\item{branch}{A character string specifying a branch name, or an object of class \dQuote{aws_repo_branch}.}

\item{name}{Optionally, a character string specifying the name for the trigger.}

\item{arn}{A character string specifying an Amazon Resource Name (ARN), such as an SNS topic.}

\item{events}{Optionally, a character vector specifying one or more events of: \dQuote{all} (the default) or any of \dQuote{updateReference} (commits are pushed), \dQuote{createReference} (new branch or tag is created), or \dQuote{deleteReference} (branch or tag is deleted).}

\item{custom_data}{Optionally, a character string specifying additional information to be passed with the trigger notice.}

\item{\dots}{Additional arguments passed to \code{\link{awscodeHTTP}}.}
}
\value{
A list.
}
\description{
Get, set, and test CodeCommit repo triggers
}

